+++ CHANGELOG.md
@@ -17,11 +17,13 @@
   - For logging purposes, the design will be scanned for submodules, ports and registers
     - The instance names of these components is then added as an attribute to each component
 - Changes to `Port`:
+  - Added **Input** and **Output** ports
   - Ports now have a **type** associated with them
     - Only values of the same type can be written to the port
     - The type is a mandatory parameter when creating a new port
     - There is runtime type checking in place, e.g., when connecting to ports, their types have to match
   - Ports now have a default value (instead of `None`)
+    - The port's type's default value is used
     - As suggested in #4, a forced propagation shall happen at the very first write of any port
   - Calling of the onchange handler is not restricted to non-root ports anymore
   - _Input_ ports now have an optional **sensitivity list**